restructure of exceptions folder and new exception
